refactor domain into a module
